# seedbox-service/openapi.yaml
openapi: 3.0.3
info:
  title: qBittorrent WebUI API
  description: |
    - All API methods are under `/api/v2/APIName/methodName`, where `APIName` is a certain subgroup of API methods whose functionality is related.
    - Either `GET` or `POST` can be used as the request type for all API methods.
    - All API methods require [authentication](https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#authentication) (except the `/api/v2/auth/login` method itself, obviously).
  termsOfService: 'https://github.com/qbittorrent/qBittorrent/blob/master/COPYING'
  license:
    name: GNU GPLv2
    url: 'https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html'
  version: 2.8.3
servers:
  - url: '/api/v2'
    description: qBittorrent Web UI API
paths:
...
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#logout'
        description: WebUI-API on qBittorrent wiki
      operationId: authLogoutPost
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
...
  /app/buildInfo:
    get:
      tags:
        - app
      summary: Get build info
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-build-info'
        description: WebUI-API on qBittorrent wiki
      operationId: appBuildInfoGet
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildInfo'
      security:
        - sid: []
  /app/shutdown:
    get:
      tags:
        - app
      summary: Shutdown application
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#shutdown-application'
        description: WebUI-API on qBittorrent wiki
      operationId: appShutdownGet
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
...
  /app/setPreferences:
    post:
      tags:
        - app
      summary: Set application preferences
      description: |
        Notes:
        1. There is no need to pass all possible preferences' `token:value` pairs if you only want to change one option
        2. Paths in `scan_dirs` must exist, otherwise this option will have no effect
        3. String values must be quoted; integer and boolean values must never be quoted

        For a list of possible preference options see [Get application preferences](https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-application-preferences)
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#set-application-preferences'
        description: WebUI-API on qBittorrent wiki
      operationId: appSetPreferencesPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                json:
                  $ref: '#/components/schemas/SetPreferences'
              required:
                - json
            example:
              json: '{"save_path":"C:/Users/Dayman/Downloads","queueing_enabled":false,"scan_dirs":{"C:/Games": 0,"D:/Downloads": 1}}'
            encoding:
              json:
                contentType: application/json
        required: true
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
  /app/defaultSavePath:
    get:
      tags:
        - app
      summary: Get default save path
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-default-save-path'
        description: WebUI-API on qBittorrent wiki
      operationId: appDefaultSavePathGet
      responses:
        '200':
          description: All scenarios
          content:
            text/plain:
              schema:
                type: string
                description: 'The response is a string with the default save path, e.g. `C:/Users/Dayman/Downloads`.'
                example: 'C:/Users/Dayman/Downloads'
      security:
        - sid: []
  /log/main:
    post:
      tags:
        - log
      summary: Get log
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-log'
        description: WebUI-API on qBittorrent wiki
      operationId: logMainPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                normal:
                  type: boolean
                  description: 'Include normal messages (default: `true`)'
                  default: true
                info:
                  type: boolean
                  description: 'Include info messages (default: `true`)'
                  default: true
                warning:
                  type: boolean
                  description: 'Include warning messages (default: `true`)'
                  default: true
                critical:
                  type: boolean
                  description: 'Include critical messages (default: `true`)'
                  default: true
                last_known_id:
                  type: integer
                  format: int64
                  description: 'Exclude messages with "message id" <= `last_known_id` (default: `-1`)'
                  default: -1
              required:
                - normal
                - info
                - warning
                - critical
                - last_known_id
            example: /api/v2/log/main?normal=true&info=true&warning=true&critical=true&last_known_id=-1
        required: true
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                type: array
                description: |
                  The response is a JSON array in which each element is an entry of the log.

                  Each element of the array has the following properties:
                items:
                  $ref: '#/components/schemas/MainLog'
              example: |
                [
                  {
                    "id":0,
                    "message":"qBittorrent v3.4.0 started",
                    "timestamp":1507969127860,
                    "type":1
                  },
                  {
                    "id":1,
                    "message":"qBittorrent is trying to listen on any interface port: 19036",
                    "timestamp":1507969127869,
                    "type":2
                  },
                  {
                    "id":2,
                    "message":"Peer ID: -qB3400-",
                    "timestamp":1507969127870,
                    "type":1
                  },
                  {
                    "id":3,
                    "message":"HTTP User-Agent is 'qBittorrent/3.4.0'",
                    "timestamp":1507969127870,
                    "type":1
                  },
                  {
                    "id":4,
                    "message":"DHT support [ON]",
                    "timestamp":1507969127871,
                    "type":2
                  },
                  {
                    "id":5,
                    "message":"Local Peer Discovery support [ON]",
                    "timestamp":1507969127871,
                    "type":2
                  },
                  {
                    "id":6,
                    "message":"PeX support [ON]",
                    "timestamp":1507969127871,
                    "type":2
                  },
                  {
                    "id":7,
                    "message":"Anonymous mode [OFF]",
                    "timestamp":1507969127871,
                    "type":2
                  },
                  {
                    "id":8,
                    "message":"Encryption support [ON]",
                    "timestamp":1507969127871,
                    "type":2
                  },
                  {
                    "id":9,
                    "message":"Embedded Tracker [OFF]",
                    "timestamp":1507969127871,
                    "type":2
                  },
                  {
                    "id":10,
                    "message":"UPnP / NAT-PMP support [ON]",
                    "timestamp":1507969127873,
                    "type":2
                  },
                  {
                    "id":11,
                    "message":"Web UI: Now listening on port 8080",
                    "timestamp":1507969127883,
                    "type":1
                  },
                  {
                    "id":12,
                    "message":"Options were saved successfully.",
                    "timestamp":1507969128055,
                    "type":1
                  },
                  {
                    "id":13,
                    "message":"qBittorrent is successfully listening on interface :: port: TCP/19036",
                    "timestamp":1507969128270,
                    "type":2
                  },
                  {
                    "id":14,
                    "message":"qBittorrent is successfully listening on interface 0.0.0.0 port: TCP/19036",
                    "timestamp":1507969128271,
                    "type":2
                  },
                  {
                    "id":15,
                    "message":"qBittorrent is successfully listening on interface 0.0.0.0 port: UDP/19036",
                    "timestamp":1507969128272,
                    "type":2
                  }
                ]
      security:
        - sid: []
  /log/peers:
    post:
      tags:
        - log
      summary: Get peer log
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-peer-log'
        description: WebUI-API on qBittorrent wiki
      operationId: logPeersPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                last_known_id:
                  type: integer
                  format: int64
                  description: 'Exclude messages with "message id" <= `last_known_id` (default: `-1`)'
                  default: -1
              required:
                - last_known_id
        required: true
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                type: array
                description: The response a JSON array. Each element of the array of objects (each object is the information relative to a peer) containing the following fields
                items:
                  $ref: '#/components/schemas/PeersLog'
      security:
        - sid: []
  /sync/maindata:
    post:
      tags:
        - sync
      summary: Get main data
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-main-data'
        description: WebUI-API on qBittorrent wiki
      operationId: syncMaindataPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                rid:
                  type: integer
                  format: int64
                  description: 'Response ID. If not provided, `rid=0` will be assumed. If the given `rid` is different from the one of last server reply, `full_update` will be `true` (see the server reply details for more info)'
            example: /api/v2/sync/maindata?rid=14
        required: true
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MainData'
              example: |
                {
                  "rid":15,
                  "torrents":
                  {
                    "8c212779b4abde7c6bc608063a0d008b7e40ce32":
                    {
                        "state":"pausedUP"
                    }
                  }
                }
      security:
        - sid: []
  /sync/torrentPeers:
    post:
      tags:
        - sync
      summary: Get torrent peers data
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-peers-data'
        description: WebUI-API on qBittorrent wiki
      operationId: syncTorrentPeersPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                hash:
                  type: string
                  description: Torrent hash
                rid:
                  type: integer
                  format: int64
                  description: 'Response ID. If not provided, `rid=0` will be assumed. If the given `rid` is different from the one of last server reply, `full_update` will be `true` (see the server reply details for more info)'
              required:
                - hash
            example: /api/v2/sync/torrentPeers?hash=8c212779b4abde7c6bc608063a0d008b7e40ce32?rid=14
        required: true
      responses:
        '200':
          description: All other scenarios- see JSON below
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TorrentPeers'
        '404':
          description: Torrent hash was not found
      security:
        - sid: []
  /transfer/info:
    get:
      tags:
        - transfer
      summary: Get global transfer info
      description: This method returns info you usually see in qBt status bar.
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-global-transfer-info'
        description: WebUI-API on qBittorrent wiki
      operationId: transferInfoGet
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferInfo'
      security:
        - sid: []
  /transfer/speedLimitsMode:
    get:
      tags:
        - transfer
      summary: Get alternative speed limits state
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-alternative-speed-limits-state'
        description: WebUI-API on qBittorrent wiki
      operationId: transferSpeedLimitsModeGet
      responses:
        '200':
          description: All scenarios
          content:
            text/plain:
              schema:
                type: integer
                format: int32
                enum:
                  - 0
                  - 1
                description: 'The response is `1` if alternative speed limits are enabled, `0` otherwise.'
      security:
        - sid: []
  /transfer/toggleSpeedLimitsMode:
    get:
      tags:
        - transfer
      summary: Toggle alternative speed limits
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#toggle-alternative-speed-limits'
        description: WebUI-API on qBittorrent wiki
      operationId: transferToggleSpeedLimitsModeGet
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
  /transfer/downloadLimit:
    get:
      tags:
        - transfer
      summary: Get global download limit
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-global-download-limit'
        description: WebUI-API on qBittorrent wiki
      operationId: transferDownloadLimitGet
      responses:
        '200':
          description: All scenarios
          content:
            text/plain:
              schema:
                type: integer
                format: int64
                description: The response is the value of current global download speed limit in bytes/second; this value will be zero if no limit is applied.
      security:
        - sid: []
  /transfer/setDownloadLimit:
    post:
      tags:
        - transfer
      summary: Set global download limit
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#set-global-download-limit'
        description: WebUI-API on qBittorrent wiki
      operationId: transferSetDownloadLimitPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  format: int64
                  description: The global download speed limit to set in bytes/second
        required: true
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
  /transfer/uploadLimit:
    get:
      tags:
        - transfer
      summary: Get global upload limit
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-global-upload-limit'
        description: WebUI-API on qBittorrent wiki
      operationId: transferUploadLimitGet
      responses:
        '200':
          description: All scenarios
          content:
            text/plain:
              schema:
                type: integer
                format: int64
                description: The response is the value of current global upload speed limit in bytes/second; this value will be zero if no limit is applied.
      security:
        - sid: []
  /transfer/setUploadLimit:
    post:
      tags:
        - transfer
      summary: Set global upload limit
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#set-global-upload-limit'
        description: WebUI-API on qBittorrent wiki
      operationId: transferSetUploadLimitPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  format: int64
                  description: The global upload speed limit to set in bytes/second
        required: true
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
  /transfer/banPeers:
    post:
      tags:
        - transfer
      summary: Ban peers
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#ban-peers'
        description: WebUI-API on qBittorrent wiki
      operationId: transferBanPeersPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                peers:
                  type: array
                  items:
                    type: string
                  description: 'The peer to ban, or multiple peers separated by a pipe `|`. Each peer is a colon-separated `host:port`'
            encoding:
              peers:
                style: pipeDelimited
                explode: false
        required: true
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
...
  /torrents/properties:
    post:
      tags:
        - torrents
      summary: Get torrent generic properties
      description: 'Requires knowing the torrent hash. You can get it from [torrent list](https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-list).'
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-generic-properties'
        description: WebUI-API on qBittorrent wiki
      operationId: torrentsPropertiesPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - hash
              properties:
                hash:
                  type: string
                  description: The hash of the torrent you want to get the generic properties of
        required: true
      responses:
        '200':
          description: All other scenarios- see JSON below
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TorrentsProperties'
              example: |
                {
                  "addition_date":1438429165,
                  "comment":"\"Debian CD from cdimage.debian.org\"",
                  "completion_date":1438429234,
                  "created_by":"",
                  "creation_date":1433605214,
                  "dl_limit":-1,
                  "dl_speed":0,
                  "dl_speed_avg":9736015,
                  "eta":8640000,
                  "last_seen":1438430354,
                  "nb_connections":3,
                  "nb_connections_limit":250,
                  "peers":1,
                  "peers_total":89,
                  "piece_size":524288,
                  "pieces_have":1254,
                  "pieces_num":1254,
                  "reannounce":672,
                  "save_path":"/Downloads/debian-8.1.0-amd64-CD-1.iso",
                  "seeding_time":1128,
                  "seeds":1,
                  "seeds_total":254,
                  "share_ratio":0.00072121022562178299,
                  "time_elapsed":1197,
                  "total_downloaded":681521119,
                  "total_downloaded_session":681521119,
                  "total_size":657457152,
                  "total_uploaded":491520,
                  "total_uploaded_session":491520,
                  "total_wasted":23481724,
                  "up_limit":-1,
                  "up_speed":0,
                  "up_speed_avg":410
                }
        '404':
          description: Torrent hash was not found
      security:
        - sid: []
  /torrents/trackers:
    post:
      tags:
        - torrents
      summary: Get torrent trackers
      description: 'Requires knowing the torrent hash. You can get it from [torrent list](https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-list).'
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-trackers'
        description: WebUI-API on qBittorrent wiki
      operationId: torrentsTrackersPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                hash:
                  type: string
                  description: The hash of the torrent you want to get the trackers of
              required:
                - hash
        required: true
      responses:
        '200':
          description: All other scenarios- see JSON below
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TorrentsTrackers'
                description: 'The response is a JSON array, where each element contains info about one tracker, with the following fields'
              example: |
                [
                  {
                  "msg":"",
                  "num_peers":100,
                  "status":2,
                  "url":"http://bttracker.debian.org:6969/announce"
                  }
                ]
        '404':
          description: Torrent hash was not found
      security:
        - sid: []
...
  /search/updatePlugins:
    post:
      tags:
        - search
      summary: Update search plugins
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#update-search-plugins'
        description: WebUI-API on qBittorrent wiki
      operationId: searchUpdatePluginsPost
      responses:
        '200':
          description: All scenarios
      security:
        - sid: []
components:
  schemas:
    BuildInfo:
      type: object
      description: The response is a JSON object containing the following fields
      properties:
        qt:
          type: string
        libtorrent:
          type: string
        boost:
          type: string
        openssl:
          type: string
        bitness:
          type: integer
          format: int32
...
    PeersLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the peer
        ip:
          type: string
          description: IP of the peer
        timestamp:
          type: integer
          format: int64
          description: Milliseconds since epoch
        blocked:
          type: boolean
          description: Whether or not the peer was blocked
        reason:
          type: string
          description: Reason of the block
...
  securitySchemes:
    sid:
      type: apiKey
      name: SID
      in: cookie
...