# seedbox-service/stubs.go
package main

import (
	"net/http"

	"github.com/labstack/echo/v4"
)

func (m *MyServer) AppBuildInfoGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) AppDefaultSavePathGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) AppDefaultSavePathPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) AppSetPreferencesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) AppShutdownGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) AppVersionGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) AuthLogoutPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) LogMainPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) LogPeersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssAddFeedPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssAddFolderPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssItemsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssMarkAsReadPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssMatchingArticlesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssMoveItemPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssRefreshItemPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssRemoveItemPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssRemoveRulePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssRenameRulePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssRulesGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) RssSetRulePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchDeletePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchEnablePluginPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchInstallPluginPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchPluginsGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchResultsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchStartPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchStatusPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchStopPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchUninstallPluginPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SearchUpdatePluginsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SyncMaindataPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) SyncTorrentPeersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsAddPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsAddPeersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsAddTagsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsAddTrackersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsBottomPrioPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsCreateCategoryPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsCreateTagsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsDecreasePrioPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsDeletePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsDeleteTagsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsDownloadLimitPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsEditCategoryPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsEditTrackerPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsFilePrioPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsFilesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsIncreasePrioPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsPausePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsPieceHashesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsPieceStatesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsPropertiesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsReannouncePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRecheckPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRemoveCategoriesPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRemoveTagsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRemoveTrackersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRenamePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRenameFilePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsRenameFolderPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsResumePost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetAutoManagementPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetCategoryPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetDownloadLimitPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetForceStartPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetLocationPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetShareLimitsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetSuperSeedingPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsSetUploadLimitPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsTagsGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsToggleFirstLastPiecePrioPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsToggleSequentialDownloadPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsTopPrioPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsTrackersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentsUploadLimitPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TorrentWebseedsPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferBanPeersPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferDownloadLimitGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferInfoGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferSetDownloadLimitPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferSetUploadLimitPost(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferSpeedLimitsModeGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferToggleSpeedLimitsModeGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}

func (m *MyServer) TransferUploadLimitGet(c echo.Context) error {
	return c.String(http.StatusOK, "Not implemented yet")
}